Restructured minigame and added a temporary FPS display
Refactored LumberingMinigame to use an init structure, the old method was killing the gameplay fps but this has fixed that. Cleaned up the mini-game while I was at it best I could for now.
Introduced FPS overlay via mixins (GuiMixin and ScreenMixin) for temporary tests on GUI and screen. Updated mixin configuration to include the new mixins.